Separate MessageLayoutManager as another file
