+++ Sources/Views/MessageLabel.swift
@@ -541,22 +541,3 @@ internal enum MessageTextCheckingType {
     case transitInfoComponents([NSTextCheckingKey: String]?)
     case custom(pattern: String, match: String?)
 }
-
-
-private class MessageLayoutManager: NSLayoutManager {
-    @available(iOS 13.0, *)
-    override func showCGGlyphs(_ glyphs: UnsafePointer<CGGlyph>,
-                               positions: UnsafePointer<CGPoint>,
-                                   count glyphCount: Int,
-                                    font: UIFont,
-                              textMatrix: CGAffineTransform,
-                              attributes: [NSAttributedString.Key : Any] = [:],
-                              in CGContext: CGContext) {
-
-        if let foregroundColor = attributes[NSAttributedString.Key.foregroundColor] as? UIColor {
-            CGContext.setFillColor(foregroundColor.cgColor)
-        }
-
-        super.showCGGlyphs(glyphs, positions: positions, count: glyphCount, font: font, textMatrix: textMatrix, attributes: attributes, in: CGContext)
-    }
-}
